move hashLinkScroll to it's own file
